make validity more robust
