+++ lumen/ai/models.py
@@ -56,12 +56,23 @@ class Sql(BaseModel):
 
 class Validity(BaseModel):
 
-    missing_table_or_columns: list[str] = Field(
-        description="List out all the missing tables or columns requested by the user."
+    chain_of_thought: str = Field(
+        description="""
+        Think about the validity of the current table.
+        Does it sufficiently include all the necessary data to answer the user's query?
+        """
+    )
+
+    missing_tables: list[str] | None = Field(
+        description="List out all the missing tables or columns requested by the user. Be mindful of name aliases."
+    )
+
+    missing_columns: list[str] | None = Field(
+        description="List out all the missing columns requested by the user."
     )
 
     is_invalid: bool = Field(
-        description="Whether the table needs a refresh."
+        description="Whether the table needs a refresh based on missing tables or columns."
     )
