+++ lumen/ai/prompts/check_validity.jinja2
@@ -1,5 +1,5 @@
 Based on the latest user's query, decide whether the
-current table and columns contain all the required info to
+current table and data contain all the required info to
 provide a reliable, accurate answer to the user's query.
 Return `is_invalid=True` if not; else return `is_invalid=False`.
 
@@ -8,7 +8,7 @@ Return `is_invalid=True` if not; else return `is_invalid=False`.
 {{ table }}
 ```
 
-### Current Columns:
+### Current Data:
 ```
-{{ columns }}
+{{ spec }}
 ```
